Multi-role usage of snap?
So I'm not quite sure how to maintain my snap packages. I don't want to
spread it all over the place, as there's a significant amount of logic
to make sure we don't try to install snap packages while snapd isn't
even running (such as in a chroot).

For now the plan is to use meta.yml, and do role dependencies, thus
allowing the `set_fact: is_snapd_running` variable to be set for the
role we're running as.

The problem with that is that the dependent role will have to make sure
that not running snapd won't cause issues for everything else.

I'm starting to lose it with snap for this computer though. So I'll
probably slowly move away from it in the next coming months.